requirements.txt
django==3.0.5
django-jchart==0.4.2
